Revert "0.4.0 rename to pyexiftool and make sure tests pass"
This reverts commit de22a9624c4ff256a05616fbded54494af86b61e.

This reverts the rename of "exiftool" to "pyexiftool" at the suggestion of the community.  The exiftool name will remain for all of eternity.
